requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
httpx==0.24.1 